Refactor SSL types to be defined in ssl_types.h
Moving SSL type definitions to ssl_types.h aims to eliminate circular dependencies when including some OpenSSL compatibility headers. One example is crypto.h, which when included alone or first errors out during compilation due to undefined types.
