+++ wolfssl/ocsp.h
@@ -28,7 +28,7 @@
 
 #ifdef HAVE_OCSP
 
-#include <wolfssl/ssl.h>
+#include <wolfssl/ssl_types.h>
 #include <wolfssl/wolfcrypt/asn.h>
 
 #ifdef __cplusplus
