+++ wolfssl/openssl/bio.h
@@ -25,7 +25,7 @@
 #ifndef WOLFSSL_BIO_H_
 #define WOLFSSL_BIO_H_
 
-#include <wolfssl/openssl/ssl.h>
+#include <wolfssl/ssl_types.h>
 
 
 #ifdef __cplusplus
