+++ wolfssl/openssl/md5.h
@@ -30,6 +30,7 @@
 #ifndef NO_MD5
 
 #include <wolfssl/wolfcrypt/hash.h>
+#include <wolfssl/ssl_types.h>
 
 #ifdef WOLFSSL_PREFIX
 #include "prefix_md5.h"
@@ -40,21 +41,11 @@
 #endif
 
 
-typedef struct WOLFSSL_MD5_CTX {
-    /* big enough to hold wolfcrypt md5, but check on init */
-#ifdef STM32_HASH
-    void* holder[(112 + WC_ASYNC_DEV_SIZE + sizeof(STM32_HASH_Context)) / sizeof(void*)];
-#else
-    void* holder[(112 + WC_ASYNC_DEV_SIZE) / sizeof(void*)];
-#endif
-} WOLFSSL_MD5_CTX;
-
 WOLFSSL_API int wolfSSL_MD5_Init(WOLFSSL_MD5_CTX*);
 WOLFSSL_API int wolfSSL_MD5_Update(WOLFSSL_MD5_CTX*, const void*, unsigned long);
 WOLFSSL_API int wolfSSL_MD5_Final(unsigned char*, WOLFSSL_MD5_CTX*);
 WOLFSSL_API int wolfSSL_MD5_Transform(WOLFSSL_MD5_CTX*, const unsigned char*);
 
-typedef WOLFSSL_MD5_CTX MD5_CTX;
 
 #define MD5_Init wolfSSL_MD5_Init
 #define MD5_Update wolfSSL_MD5_Update
