+++ wolfssl/openssl/objects.h
@@ -24,11 +24,7 @@
 #define WOLFSSL_OBJECTS_H_
 
 #include <wolfssl/wolfcrypt/settings.h>
-//#include <wolfssl/openssl/ssl.h>
-#ifndef OPENSSL_EXTRA_SSL_GUARD
-#define OPENSSL_EXTRA_SSL_GUARD
-#include <wolfssl/ssl.h>
-#endif /* OPENSSL_EXTRA_SSL_GUARD */
+#include <wolfssl/ssl_types.h>
 
 #ifdef __cplusplus
     extern "C" {
