+++ wolfssl/openssl/pem.h
@@ -33,7 +33,7 @@
 #include <wolfssl/openssl/bio.h>
 #include <wolfssl/openssl/rsa.h>
 #include <wolfssl/openssl/dsa.h>
-#include <wolfssl/ssl.h>
+#include <wolfssl/ssl_types.h>
 
 #ifdef __cplusplus
     extern "C" {
